Move the basic implementation of FDTD field solver to the new eponymous template
Now Yee solver refers to only the classic Yee solver.
The general solver is FDTD<CurlE, CurlB>.
Update comments to reflect it.
In the fieldSolver.param, the Yee solver became just Yee, not Yee<>.
