+++ include/picongpu/fields/MaxwellSolver/ArbitraryOrderFDTD/ArbitraryOrderFDTD.def
@@ -20,7 +20,7 @@
 #pragma once
 
 #include "picongpu/fields/MaxwellSolver/ArbitraryOrderFDTD/Derivative.def"
-#include "picongpu/fields/MaxwellSolver/Yee/Yee.def"
+#include "picongpu/fields/MaxwellSolver/FDTD/FDTD.def"
 #include "picongpu/fields/differentiation/Curl.def"
 
 #include <cstdint>
@@ -70,9 +70,9 @@ namespace picongpu
              */
             template<uint32_t T_neighbors>
             using ArbitraryOrderFDTD
-                = ::picongpu::fields::maxwellSolver::Yee<aoFDTD::CurlE<T_neighbors>, aoFDTD::CurlB<T_neighbors>>;
+                = ::picongpu::fields::maxwellSolver::FDTD<aoFDTD::CurlE<T_neighbors>, aoFDTD::CurlB<T_neighbors>>;
 
-            /* We need no definition of margins, because the Yee solver uses its curl
+            /* We need no definition of margins, because the FDTD solver uses its curl
              * classes to define margins
              */
