+++ include/picongpu/fields/MaxwellSolver/Yee/Yee.hpp
@@ -22,21 +22,10 @@
 
 #include "picongpu/simulation_defines.hpp"
 
-#include "picongpu/fields/FieldB.hpp"
-#include "picongpu/fields/FieldE.hpp"
-#include "picongpu/fields/LaserPhysics.hpp"
 #include "picongpu/fields/MaxwellSolver/CFLChecker.hpp"
-#include "picongpu/fields/MaxwellSolver/Yee/Yee.kernel"
-#include "picongpu/fields/absorber/Absorber.hpp"
-#include "picongpu/fields/absorber/pml/Pml.hpp"
-#include "picongpu/fields/cellType/Yee.hpp"
-#include "picongpu/fields/incidentField/Solver.hpp"
-#include "picongpu/traits/GetMargin.hpp"
+#include "picongpu/fields/MaxwellSolver/Yee/Yee.def"
 
 #include <pmacc/traits/GetStringProperties.hpp>
-
-#include <memory>
-#include <stdexcept>
 
 
 namespace picongpu
@@ -45,242 +34,9 @@ namespace picongpu
     {
         namespace maxwellSolver
         {
-            /** Finite-difference time-domain (FDTD) Maxwell's solver on a regular Yee grid
-             *
-             * This class template implements FDTD scheme for any given curl types.
-             * All supported field solvers, except placeholder None, are aliases of this template.
-             *
-             * @tparam T_CurlE functor to compute curl of E
-             * @tparam T_CurlB functor to compute curl of B
-             */
-            template<typename T_CurlE, typename T_CurlB>
-            class Yee
-            {
-            public:
-                // Types required by field solver interface
-                using CellType = cellType::Yee;
-                using CurlE = T_CurlE;
-                using CurlB = T_CurlB;
-
-                Yee(MappingDesc const cellDescription) : cellDescription(cellDescription)
-                {
-                    CFLChecker<Yee>{}();
-                    DataConnector& dc = Environment<>::get().DataConnector();
-                    fieldE = dc.get<FieldE>(FieldE::getName(), true);
-                    fieldB = dc.get<FieldB>(FieldB::getName(), true);
-                    auto& absorberFactory = fields::absorber::AbsorberFactory::get();
-                    absorberImpl = absorberFactory.makeImpl(cellDescription);
-                }
-
-                /** Perform the first part of E and B propagation by a time step.
-                 *
-                 * Together with update_afterCurrent( ) forms the full propagation.
-                 *
-                 * @param currentStep index of the current time iteration
-                 */
-                void update_beforeCurrent(uint32_t const currentStep)
-                {
-                    /* As typical for electrodynamic PIC codes, we split an FDTD update of B into two halves.
-                     * (This comes from commonly used particle pushers needing E and B at the same time.)
-                     * Here we do the second half of updating B.
-                     * It completes the first half in update_afterCurrent() at the previous time step.
-                     * In vacuum, the updates are linear and splitting could only influence floating-point arithmetic.
-                     * For PML there is a difference, it is treated inside the absorber implementation.
-                     * In both cases, the full E and B fields behave as expected after the update.
-                     */
-                    updateBSecondHalf<CORE + BORDER>(currentStep);
-                    auto incidentFieldSolver = fields::incidentField::Solver{cellDescription};
-                    // update B by half step, to step = currentStep + 0.5, so step for E_inc = currentStep
-                    incidentFieldSolver.updateBHalf(static_cast<float_X>(currentStep));
-                    EventTask eRfieldB = fieldB->asyncCommunication(__getTransactionEvent());
-
-                    updateE<CORE>(currentStep);
-                    // Incident field solver update does not use exchanged B, so does not have to wait for it
-                    // update E by full step, to step = currentStep + 1, so step for B_inc = currentStep + 0.5
-                    incidentFieldSolver.updateE(static_cast<float_X>(currentStep) + 0.5_X);
-                    __setTransactionEvent(eRfieldB);
-                    updateE<BORDER>(currentStep);
-                }
-
-                /** Perform the last part of E and B propagation by a time step
-                 *
-                 * Together with update_beforeCurrent( ) forms the full propagation.
-                 *
-                 * @param currentStep index of the current time iteration
-                 */
-                void update_afterCurrent(uint32_t const currentStep)
-                {
-                    auto& absorber = absorber::Absorber::get();
-                    if(absorber.getKind() == absorber::Absorber::Kind::Exponential)
-                    {
-                        auto& exponentialImpl = absorberImpl->asExponentialImpl();
-                        exponentialImpl.run(currentStep, fieldE->getDeviceDataBox());
-                    }
-                    if(laserProfiles::Selected::INIT_TIME > 0.0_X)
-                        LaserPhysics{}(currentStep);
-
-                    // Incident field solver update does not use exchanged E, so does not have to wait for it
-                    auto incidentFieldSolver = fields::incidentField::Solver{cellDescription};
-                    // update B by half step, to step currentStep + 1.5, so step for E_inc = currentStep + 1
-                    incidentFieldSolver.updateBHalf(static_cast<float_X>(currentStep) + 1.0_X);
-
-                    EventTask eRfieldE = fieldE->asyncCommunication(__getTransactionEvent());
-
-                    // First and second halves of B update are explained inside update_beforeCurrent()
-                    updateBFirstHalf<CORE>(currentStep);
-                    __setTransactionEvent(eRfieldE);
-                    updateBFirstHalf<BORDER>(currentStep);
-
-                    if(absorber.getKind() == absorber::Absorber::Kind::Exponential)
-                    {
-                        auto& exponentialImpl = absorberImpl->asExponentialImpl();
-                        exponentialImpl.run(currentStep, fieldB->getDeviceDataBox());
-                    }
-
-                    EventTask eRfieldB = fieldB->asyncCommunication(__getTransactionEvent());
-                    __setTransactionEvent(eRfieldB);
-                }
-
-                static pmacc::traits::StringProperty getStringProperties()
-                {
-                    pmacc::traits::StringProperty propList("name", "Yee");
-                    return propList;
-                }
-
-            private:
-                // Helper types for configuring kernels
-                template<typename T_Curl>
-                using BlockDescription = pmacc::SuperCellDescription<
-                    SuperCellSize,
-                    typename traits::GetLowerMargin<T_Curl>::type,
-                    typename traits::GetUpperMargin<T_Curl>::type>;
-                template<uint32_t T_Area>
-                using AreaMapper = pmacc::AreaMapping<T_Area, MappingDesc>;
-
-                /** Propagate B values in the given area by the first half of a time step
-                 *
-                 * This operation propagates grid values of field B by dt/2.
-                 * If PML is used, it also prepares the internal state of convolutional components
-                 * so that calling updateBSecondHalf() afterwards competes the update.
-                 *
-                 * @tparam T_Area area to apply updates to, the curl must be applicable to all points;
-                 *                normally CORE, BORDER, or CORE + BORDER
-                 *
-                 * @param currentStep index of the current time iteration
-                 */
-                template<uint32_t T_Area>
-                void updateBFirstHalf(uint32_t const currentStep)
-                {
-                    updateBHalf<T_Area>(currentStep, true);
-                }
-
-                /** Propagate B values in the given area by the second half of a time step
-                 *
-                 * This operation propagates grid values of field B by dt/2.
-                 * If PML is used, it relies on the internal state of convolutional components
-                 * having been set up by a prior call to updateBFirstHalf().
-                 * Then this call leaves it ready for updateBFirstHalf() called on the next time step.
-                 *
-                 * @tparam T_Area area to apply updates to, the curl must be applicable to all points;
-                 *                normally CORE, BORDER, or CORE + BORDER
-                 *
-                 * @param currentStep index of the current time iteration
-                 */
-                template<uint32_t T_Area>
-                void updateBSecondHalf(uint32_t const currentStep)
-                {
-                    updateBHalf<T_Area>(currentStep, false);
-                }
-
-                /** Propagate B values in the given area by half a time step
-                 *
-                 * @tparam T_Area area to apply updates to, the curl must be applicable to all points;
-                 *                normally CORE, BORDER, or CORE + BORDER
-                 *
-                 * @param currentStep index of the current time iteration
-                 * @param updatePsiB whether convolutional magnetic fields need to be updated, or are up-to-date
-                 */
-                template<uint32_t T_Area>
-                void updateBHalf(uint32_t const currentStep, bool const updatePsiB)
-                {
-                    constexpr auto numWorkers = getNumWorkers();
-                    using Kernel = yee::KernelUpdateB<numWorkers, BlockDescription<CurlE>>;
-                    AreaMapper<T_Area> mapper{cellDescription};
-
-                    // The ugly transition from run-time to compile-time polymorphism is contained here
-                    auto& absorber = absorber::Absorber::get();
-                    if(absorber.getKind() == absorber::Absorber::Kind::Pml)
-                    {
-                        auto& pmlImpl = absorberImpl->asPmlImpl();
-                        auto const updateFunctor
-                            = pmlImpl.getUpdateBHalfFunctor<CurlE, T_Area>(currentStep, updatePsiB);
-                        PMACC_KERNEL(Kernel{})
-                        (mapper.getGridDim(),
-                         numWorkers)(mapper, updateFunctor, fieldE->getDeviceDataBox(), fieldB->getDeviceDataBox());
-                    }
-                    else
-                    {
-                        PMACC_KERNEL(Kernel{})
-                        (mapper.getGridDim(), numWorkers)(
-                            mapper,
-                            yee::UpdateBHalfFunctor<CurlE>{},
-                            fieldE->getDeviceDataBox(),
-                            fieldB->getDeviceDataBox());
-                    }
-                }
-
-                /** Propagate E values in the given area by a time step.
-                 *
-                 * @tparam T_Area area to apply updates to, the curl must be applicable to all points;
-                 *                normally CORE, BORDER, or CORE + BORDER
-                 *
-                 * @param currentStep index of the current time iteration
-                 */
-                template<uint32_t T_Area>
-                void updateE(uint32_t currentStep)
-                {
-                    constexpr auto numWorkers = getNumWorkers();
-                    using Kernel = yee::KernelUpdateE<numWorkers, BlockDescription<CurlB>>;
-                    auto mapper = AreaMapper<T_Area>{cellDescription};
-
-                    // The ugly transition from run-time to compile-time polymorphism is contained here
-                    auto& absorber = absorber::Absorber::get();
-                    if(absorber.getKind() == absorber::Absorber::Kind::Pml)
-                    {
-                        auto& pmlImpl = absorberImpl->asPmlImpl();
-                        auto const updateFunctor = pmlImpl.getUpdateEFunctor<CurlB, T_Area>(currentStep);
-                        PMACC_KERNEL(Kernel{})
-                        (mapper.getGridDim(),
-                         numWorkers)(mapper, updateFunctor, fieldB->getDeviceDataBox(), fieldE->getDeviceDataBox());
-                    }
-                    else
-                    {
-                        PMACC_KERNEL(Kernel{})
-                        (mapper.getGridDim(), numWorkers)(
-                            mapper,
-                            yee::UpdateEFunctor<CurlB>{},
-                            fieldB->getDeviceDataBox(),
-                            fieldE->getDeviceDataBox());
-                    }
-                }
-
-                //! Get number of workers for kernels
-                static constexpr uint32_t getNumWorkers()
-                {
-                    return pmacc::traits::GetNumWorkers<pmacc::math::CT::volume<SuperCellSize>::type::value>::value;
-                }
-
-                MappingDesc const cellDescription;
-                std::shared_ptr<FieldE> fieldE;
-                std::shared_ptr<FieldB> fieldB;
-
-                // Absorber implementation
-                std::unique_ptr<fields::absorber::AbsorberImpl> absorberImpl;
-            };
-
             //! Specialization of the CFL condition checker for the classic Yee solver
             template<>
-            struct CFLChecker<Yee<>>
+            struct CFLChecker<Yee>
             {
                 //! Check the CFL condition, doesn't compile when failed
                 void operator()() const
@@ -289,58 +45,29 @@ namespace picongpu
                     PMACC_CASSERT_MSG(
                         Courant_Friedrichs_Lewy_condition_failure____check_your_grid_param_file,
                         (SPEED_OF_LIGHT * SPEED_OF_LIGHT * DELTA_T * DELTA_T * INV_CELL2_SUM) <= 1.0
-                            && sizeof(Yee<>*) != 0);
+                            && sizeof(Yee*) != 0);
                 }
             };
 
         } // namespace maxwellSolver
     } // namespace fields
+} // namespace picongpu
 
+namespace pmacc
+{
     namespace traits
     {
-        /** Get margin for B field access in the Yee solver
-         *
-         * @tparam T_CurlE functor to compute curl of E
-         * @tparam T_CurlB functor to compute curl of B
-         */
-        template<typename T_CurlE, typename T_CurlB>
-        struct GetMargin<fields::maxwellSolver::Yee<T_CurlE, T_CurlB>, FieldB>
+        template<>
+        struct StringProperties<::picongpu::fields::maxwellSolver::Yee>
         {
-            using LowerMargin = typename T_CurlB::LowerMargin;
-            using UpperMargin = typename T_CurlB::UpperMargin;
-        };
-
-        /** Get margin for E field access in the Yee solver
-         *
-         * @tparam T_CurlE functor to compute curl of E
-         * @tparam T_CurlB functor to compute curl of B
-         */
-        template<typename T_CurlE, typename T_CurlB>
-        struct GetMargin<fields::maxwellSolver::Yee<T_CurlE, T_CurlB>, FieldE>
-        {
-            using LowerMargin = typename T_CurlE::LowerMargin;
-            using UpperMargin = typename T_CurlE::UpperMargin;
-        };
-
-        /** Get margin for both fields access in the Yee solver
-         *
-         * @tparam T_CurlE functor to compute curl of E
-         * @tparam T_CurlB functor to compute curl of B
-         */
-        template<typename T_CurlE, typename T_CurlB>
-        struct GetMargin<fields::maxwellSolver::Yee<T_CurlE, T_CurlB>>
-        {
-        private:
-            using Solver = fields::maxwellSolver::Yee<T_CurlE, T_CurlB>;
-
-        public:
-            using LowerMargin = typename pmacc::math::CT::max<
-                typename GetLowerMargin<Solver, FieldB>::type,
-                typename GetLowerMargin<Solver, FieldE>::type>::type;
-            using UpperMargin = typename pmacc::math::CT::max<
-                typename GetUpperMargin<Solver, FieldB>::type,
-                typename GetUpperMargin<Solver, FieldE>::type>::type;
+            static StringProperty get()
+            {
+                auto propList = ::picongpu::fields::maxwellSolver::Yee::getStringProperties();
+                // overwrite the name of the solver (inherit all other properties)
+                propList["name"].value = "Yee";
+                return propList;
+            }
         };
 
     } // namespace traits
-} // namespace picongpu
+} // namespace pmacc
