+++ share/picongpu/tests/CollisionsBeamRelaxation/include/picongpu/param/fieldSolver.param
@@ -41,7 +41,7 @@ namespace picongpu
         /** FieldSolver
          *
          * Field Solver Selection (note <> for some solvers):
-         *  - Yee<> : Standard Yee solver approximating derivatives with respect to time and
+         *  - Yee: Standard Yee solver approximating derivatives with respect to time and
          * space by second order finite differences.
          *  - Lehe<>: Num. Cherenkov free field solver in a chosen direction
          *  - ArbitraryOrderFDTD<4>: Solver using 4 neighbors to each direction to approximate
